WFC_Testing/requirements.txt
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
docopt==0.6.2
exceptiongroup==1.0.4
fonttools==4.38.0
idna==3.4
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pipreq==0.4
pipreqs==0.4.11
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
requests==2.28.1
scipy==1.9.3
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.13
yarg==0.1.9
